+++ pkg/app/controller.go
@@ -9,6 +9,7 @@ import (
 	"github.com/Octops/gameserver-ingress-controller/pkg/handlers"
 	"github.com/Octops/gameserver-ingress-controller/pkg/k8sutil"
 	"github.com/Octops/gameserver-ingress-controller/pkg/manager"
+	"github.com/Octops/gameserver-ingress-controller/pkg/record"
 	"github.com/Octops/gameserver-ingress-controller/pkg/stores"
 	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
@@ -54,7 +55,9 @@ func StartController(ctx context.Context, config Config) error {
 	}
 
 	logger.WithField("component", "controller").Info("starting gameserver controller")
-	handler := handlers.NewGameSeverEventHandler(store, mgr.GetEventRecorderFor("gameserver-ingress-controller"))
+
+	recorder := mgr.GetEventRecorderFor("gameserver-ingress-controller")
+	handler := handlers.NewGameSeverEventHandler(store, record.NewEventRecorder(recorder))
 	ctrl, err := controller.NewGameServerController(mgr, handler, controller.Options{
 		For: &agonesv1.GameServer{},
 	})
