+++ pkg/reconcilers/service_reconciler.go
@@ -5,12 +5,12 @@ import (
 	"context"
 	"github.com/Octops/gameserver-ingress-controller/internal/runtime"
 	"github.com/Octops/gameserver-ingress-controller/pkg/gameserver"
+	"github.com/Octops/gameserver-ingress-controller/pkg/record"
 	"github.com/pkg/errors"
 	corev1 "k8s.io/api/core/v1"
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/intstr"
-	"k8s.io/client-go/tools/record"
 )
 
 type ServiceStore interface {
@@ -20,13 +20,13 @@ type ServiceStore interface {
 
 type ServiceReconciler struct {
 	store    ServiceStore
-	recorder *EventRecorder
+	recorder *record.EventRecorder
 }
 
-func NewServiceReconciler(store ServiceStore, recorder record.EventRecorder) *ServiceReconciler {
+func NewServiceReconciler(store ServiceStore, recorder *record.EventRecorder) *ServiceReconciler {
 	return &ServiceReconciler{
 		store:    store,
-		recorder: NewEventRecorder(recorder),
+		recorder: recorder,
 	}
 }
 
@@ -45,24 +45,24 @@ func (r *ServiceReconciler) Reconcile(ctx context.Context, gs *agonesv1.GameServ
 }
 
 func (r *ServiceReconciler) reconcileNotFound(ctx context.Context, gs *agonesv1.GameServer) (*corev1.Service, error) {
-	r.recorder.RecordCreating(gs, ServiceKind)
+	r.recorder.RecordCreating(gs, record.ServiceKind)
 
 	service, err := newService(gs)
 	if err != nil {
-		r.recorder.RecordFailed(gs, ServiceKind, err)
+		r.recorder.RecordFailed(gs, record.ServiceKind, err)
 		return nil, errors.Wrapf(err, "failed to create service for gameserver %s", gs.Name)
 	}
 
 	result, err := r.store.CreateService(ctx, service, metav1.CreateOptions{})
 	if err != nil {
 		if !k8serrors.IsAlreadyExists(err) {
-			r.recorder.RecordFailed(gs, ServiceKind, err)
+			r.recorder.RecordFailed(gs, record.ServiceKind, err)
 			return nil, errors.Wrap(err, "failed to create service")
 		}
 		runtime.Logger().Debug(err)
 	}
 
-	r.recorder.RecordSuccess(gs, ServiceKind)
+	r.recorder.RecordSuccess(gs, record.ServiceKind)
 	return result, nil
 }
